bug fix: certificate generator image rendering problem
feat: add Vivaldi font and certificate design to ticket generator page
The Vivaldi font is added to the project and used in the certificate design. The certificate design is updated to use html2canvas to generate a PNG image of the certificate with the attendee's name. The previous implementation used an API endpoint to generate the certificate image. The new implementation is more efficient and allows for more flexibility in the design of the certificate. The CSS styles for the certificate design are added to the ticket generator page.